backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
django-allauth==0.63.3
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
drfspectacular==0.27.2
djangorestframework-simplejwt==5.3.1
psycopg[binary]==3.2.3
Pillow==10.4.0
django-storages==1.14.3

